pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tagdiff"
version = "0.1.1"
description = "Compare GitHub release changelogs between tags"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "ar9av", email = "ar9avg@gmail.com"},
]
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "requests>=2.31.0",
  "litellm>=1.0.0",
  "python-dotenv>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/ar9av/tagdiff"
"Bug Tracker" = "https://github.com/ar9av/tagdiff/issues"

[project.scripts]
tagdiff = "tagdiff.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tagdiff*"]
